fixes to download users csv
